backend/requirements.txt
Django==4.1.9
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2